Use  http server for C struct -> asciiTable conversion
