+++ main.go
@@ -1,55 +1,45 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
-	"github.com/olekukonko/tablewriter"
-	"log"
-	"os"
+	"github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin/binding"
+	"net/http"
 )
 
-func SetTableFormat(table *tablewriter.Table) {
-	table.SetAlignment(tablewriter.ALIGN_LEFT) // Set Alignment
-	table.SetAutoMergeCells(true)
-	table.SetRowLine(true)
-	table.SetAlignment(tablewriter.ALIGN_CENTER)
-	table.SetHeader([]string{"TypeName", "VarName", "Extra"})
-}
-
-func getTable(c *CVar, tables map[string][][]string) {
-	var t [][]string
-	t = append(t, []string{c.getTypeName(), c.getVarName(), c.Comment})
-	for _, cvar := range c.CVarList {
-		t = append(t, []string{cvar.getTypeName(), cvar.getVarName(), cvar.Comment})
-		if cvar.CVarList != nil {
-			getTable(cvar, tables)
-		}
-	}
-	tables[c.TypeName] = t
-}
-
-func getTableFormatString(data [][]string) string {
-	var writer bytes.Buffer
-	table := tablewriter.NewWriter(&writer)
-	SetTableFormat(table)
-	table.AppendBulk(data)
-	table.Render()
-	return writer.String()
+type ParseBody struct {
+	Language string `json:"language"`
+	Body     string `json:"body"`
 }
 
 func main() {
-	s := make(map[string][][]string)
-	var c CVar
-	if len(os.Args) < 2 {
-		log.Fatal("args must Greater than 1")
-	}
-	//cStr := os.Args[1]
 
-	cStr := os.Args[1]
-	c.parse(cStr)
-	getTable(&c, s)
-	for _, v := range s {
-		fmt.Printf("%s", MultilineComment(getTableFormatString(v)))
-		fmt.Printf("\n\n\n")
-	}
+	r := gin.Default()
+	r.GET("/ping", func(context *gin.Context) {
+		context.JSON(http.StatusOK, gin.H{
+			"message": "pong",
+		})
+	})
+
+	r.POST("/ascii-table", func(context *gin.Context) {
+		var body ParseBody
+
+		err := context.BindJSON(&body)
+		if err != nil {
+			return
+		}
+
+		s := make(map[string][][]string)
+		var c CVar
+		cStr := body.Body
+		c.parse(cStr)
+		getTable(&c, s)
+		var Ascii string
+		for _, v := range s {
+			Ascii += fmt.Sprintf("%s\n\n\n", MultilineComment(getTableFormatString(v)))
+		}
+		context.Data(http.StatusOK, binding.MIMEHTML, []byte(Ascii))
+	})
+	r.Run() // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
+
 }
